Makefile, libft.h, vérif

+++ Day1/ft_memccpy.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void		*ft_memccpy(void *dest, void *src, int c, unsigned int n)
+#include "libft.h"
+
+void		*ft_memccpy(void *dest, void *src, int c, size_t n)
 {
 	char	*dest2;
 	char	*src2;
